pyproject.toml
[tool.poetry]
name = "thesis"
version = "0.1.0"
description = ""
authors = ["Sameer Tantry <sameertantry@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
albumentations = "^1.4.1"
tqdm = "^4.66.2"
omegaconf = "^2.3.0"
torch = "^2.2.2"
torchvision = "^0.17.2"
lightning = "^2.2.1"
tensorboard = "^2.16.2"
hydra-core = "^1.3.2"
timm = "^0.9.16"


[tool.poetry.group.dev.dependencies]
numpy = "^1.26.4"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
pandas = "^2.2.1"
omegaconf = "^2.3.0"
opencv-python = "^4.9.0.80"
pillow = "^10.2.0"
hydra-core = "^1.3.2"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
isort = "^5.13.2"
black = "^24.2.0"
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
